+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 * Adds Tensored Type
+* Adds Module instances
 
 ## 0.1.1 -- 2021-12-13